@@ -9,14 +9,39 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "react": "^18",
+    "@radix-ui/react-dialog": "^1.1.1",
+    "@radix-ui/react-scroll-area": "^1.1.0",
+    "@radix-ui/react-select": "^2.1.1",
+    "@radix-ui/react-slot": "^1.1.0",
+    "@radix-ui/react-tabs": "^1.1.0",
+    "@radix-ui/react-toast": "^1.2.2",
+    "@radix-ui/react-tooltip": "^1.1.2",
+    "@reduxjs/toolkit": "^2.2.7",
+    "axios": "^1.7.7",
+    "class-variance-authority": "^0.7.0",
+    "clsx": "^2.1.1",
+    "countup": "^1.8.2",
+    "framer-motion": "^11.5.6",
+    "jwt-decode": "^4.0.0",
+    "lucide-react": "^0.445.0",
+    "next": "14.2.13",
+    "react": "^18.3.1",
+    "react-countup": "^6.5.3",
     "react-dom": "^18",
-    "next": "14.2.13"
+    "react-icons": "^5.3.0",
+    "react-redux": "^9.1.2",
+    "swiper": "^11.1.14",
+    "tailwind-merge": "^2.5.2",
+    "tailwindcss-animate": "^1.0.7",
+    "zod": "^3.23.8"
   },
   "devDependencies": {
-    "postcss": "^8",
-    "tailwindcss": "^3.4.1",
+    "@types/node": "22.7.4",
+    "@types/react": "18.3.11",
+    "autoprefixer": "^10.4.20",
     "eslint": "^8",
-    "eslint-config-next": "14.2.13"
+    "eslint-config-next": "14.2.13",
+    "postcss": "^8.4.47",
+    "tailwindcss": "^3.4.13"
   }
 }
